requirements.txt
# Core ML and Computer Vision
ultralytics>=8.0.0          # YOLO11 model training and inference
torch>=2.0.0                # PyTorch deep learning framework
torchvision>=0.15.0         # Computer vision utilities for PyTorch
opencv-python>=4.6.0        # Computer vision and image processing
pillow>=9.0.0               # Image processing library
numpy>=1.21.0               # Numerical computing
scipy>=1.7.0                # Scientific computing

# OCR and Text Processing
paddlepaddle>=2.4.0         # PaddlePaddle deep learning framework
paddleocr>=2.6.0            # PaddleOCR text recognition
transformers>=4.20.0        # Hugging Face transformers for LayoutLM

# PDF Processing
pdf2image>=1.16.0           # Convert PDF to images (requires poppler-utils system package)
reportlab>=3.6.0            # PDF generation and manipulation

# Data Processing and Utilities
pandas>=1.3.0               # Data manipulation and analysis
matplotlib>=3.5.0           # Plotting and visualization
tqdm>=4.64.0                # Progress bars

# Configuration and Serialization
ruamel.yaml>=0.17.0         # YAML configuration file handling
pyyaml>=6.0                 # Alternative YAML parser
json5>=0.9.0                # Extended JSON format support

# Web Interface and Visualization
streamlit>=1.20.0           # Web interface for pipeline review
plotly>=5.0.0               # Interactive plotting

# Deployment and Optimization (Optional)
onnxruntime>=1.12.0         # ONNX runtime for model optimization

# Network and Communication
grpcio>=1.71.0,<2.0.0       # gRPC communication
requests>=2.25.0            # HTTP requests
urllib3>=1.26.0             # HTTP client

# System and Process Management
psutil>=5.8.0               # System and process utilities
cloudpickle>=3.0.0,<4.0.0   # Enhanced pickling for distributed computing

# Development and Testing
pytest>=7.0.0               # Testing framework
black>=22.0.0               # Code formatting
flake8>=4.0.0               # Code linting

# Optional Web Framework (choose one)
# django>=4.0.0             # Full-featured web framework
# flask>=2.0.0              # Lightweight web framework